requirements.txt
boto3===1.4.4
botocore

ptpython
autopep8===1.3.1
pylint===1.6.5
flake8===3.3.0
pydocstyle===1.1.1
PyYAML===3.12
docopt==0.6.2
tornado==4.5.1
